requirements.txt
pymongo~=4.6.1
python-dotenv~=1.0.1
prettytable~=3.9.0